requirements-test.txt
pytest<8.0.0 # https://github.com/pytest-dev/pytest/issues/11868
pytest-cov
pytest-timeout
pytest-asyncio
respx
black
mypy
types-setuptools
pbr
time_machine<2.14.0
pre-commit
backports.zoneinfo;python_version<"3.9"
ruff==0.2.1
